ncurses 5.9 - patch 20110514
+ compiler-warning fixes in panel and progs.
+ modify CF_PKG_CONFIG macro, from changes to tin -TD
+ modify CF_CURSES_FUNCS configure macro, used in test directory
  configure script:
  + work around (non-optimizer) bug in gcc 4.2.1 which caused
    test-expression to be omitted from executable.
  + force the linker to see a link-time expression of a symbol, to
    help work around weak-symbol issues.

+++ test/color_name.h
@@ -26,7 +26,7 @@
  * authorization.                                                           *
  ****************************************************************************/
 /*
- * $Id: color_name.h,v 1.2 2011/04/23 17:15:01 tom Exp $
+ * $Id: color_name.h,v 1.3 2011/05/14 17:41:17 tom Exp $
  */
 
 #ifndef __COLORNAME_H
@@ -64,7 +64,7 @@ color_code(const char *color)
     char *endp = 0;
     size_t n;
 
-    if ((result = strtol(color, &endp, 0)) >= 0
+    if ((result = (int) strtol(color, &endp, 0)) >= 0
 	&& (endp == 0 || *endp == 0)) {
 	;
     } else if (!strcmp(color, "default")) {
